moved Page & PageRecorder into separate file
